+++ concepts/external-crates/.meta/config.json
@@ -1,0 +1,8 @@
+{
+  "blurb": "External crates: a way to prevent reinventing the wheel",
+  "authors": [
+    "still-flow",
+    "coriolinus"
+  ],
+  "contributors": []
+}
