+++ concepts/external-crates/about.md
@@ -1,0 +1,5 @@
+# About
+
+A crate is a compilation unit in Rust. A crate can be compiled into a binary or into a library. Most of the time, you will be using `cargo` to manage external dependencies, which is as simple as adding a line to your `Cargo.toml` file.
+
+- [introduction to packages and Cargo in the Rust book](https://doc.rust-lang.org/book/ch07-00-managing-growing-projects-with-packages-crates-and-modules.html)
