+++ exercises/concept/resistor-color/.docs/hints.md
@@ -1,0 +1,1 @@
+Use the `enum_iterator` crate to be able to convert integers into enums. Check out [docs](https://docs.rs/enum-iterator/0.7.0/enum_iterator/) to find out how.
