+++ exercises/concept/resistor-color/.gitignore
