--- exercises/practice/resistor-color/.docs/hints.md
@@ -1,1 +1,0 @@
-Use the `int-enum` crate to be able to convert integers into enums. Check out [docs](https://docs.rs/int-enum/0.4.0/int_enum/) to find out how.
